requirements.txt
streamlit
groq
python-dotenv
httpx
requests
pandas
openpyxl
serpapi